# docs/swagger.yaml
definitions:
  models.CommonRes:
    properties:
      error_code:
        type: integer
      message:
        type: string
    type: object
...
  models.MonitoringNginxParams:
    properties:
      common:
        $ref: '#/definitions/models.monitoringNginxCommon'
      nginx:
        $ref: '#/definitions/models.monitoringNginx'
      rtmp2rtp:
        $ref: '#/definitions/models.monitoringRtmp'
      server:
        type: string
    type: object
...
  models.monitoringNginx:
    properties:
      conf_count:
        type: integer
      health_check:
        type: boolean
    type: object
  models.monitoringNginxCommon:
    properties:
      cpu_num:
        type: integer
      cpu_used:
        description: percent
        type: integer
      memory_free:
        description: bytes
        type: integer
      memory_total:
        description: bytes
        type: integer
    type: object
  models.monitoringRtmp:
    properties:
      module_count:
        type: integer
      modules:
        items:
          $ref: '#/definitions/models.monitoringRtmpModules'
        type: array
    type: object
  models.monitoringRtmpModules:
    properties:
      channel:
        type: string
      targets:
        items:
          $ref: '#/definitions/models.monitoringRtmpModulesTargets'
        type: array
    type: object
  models.monitoringRtmpModulesTargets:
    properties:
      audio_port:
        type: integer
      ip:
        type: string
      video_port1:
        type: integer
      video_port2:
        type: integer
      video_port3:
        type: integer
    type: object
info:
  contact:
    email: deuksoo.mun@catenoid.net
    name: API Support
  description: Webrtc Controller Api
  title: WebRtc Controller Swagger API
  version: "1.0"
paths:
...
  /janusinfo:
    get:
      consumes:
      - application/json
      description: Get to Janus Agent and Janus Turn Status Monitoring information
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: object
            type: array
        "400":
          description: Invalid Parameters
          schema:
            $ref: '#/definitions/models.CommonRes'
        "401":
          description: Validation Authorization
          schema:
            $ref: '#/definitions/models.CommonRes'
        "417":
          description: Janus Monitoring Info got Exception
          schema:
            $ref: '#/definitions/models.CommonRes'
      summary: Get Janus Monitoring information
      tags:
      - Monitoring
    post:
      consumes:
      - application/json
      description: Janus Agent and Janus Turn Status Monitoring information register
      parameters:
      - description: BroadCast is close. All Agent and live Stream Api send to broadcast
          close signal
        in: body
        name: BroadCastCloseInfo
        required: true
        schema:
          $ref: '#/definitions/models.CommonRes'
      produces:
      - application/json
      responses:
        "200":
          description: Janus Monitoring Info registered
          schema:
            $ref: '#/definitions/models.CommonRes'
        "400":
          description: Invalid Parameters
          schema:
            $ref: '#/definitions/models.CommonRes'
        "401":
          description: Validation Authorization
          schema:
            $ref: '#/definitions/models.CommonRes'
        "417":
          description: Janus Monitoring Info registered Exception
          schema:
            $ref: '#/definitions/models.CommonRes'
      summary: Register Janus Monitoring information
      tags:
      - Monitoring
  /nginxinfo:
    get:
      consumes:
      - application/json
      description: Get to Nginx Agent and Nginx Turn Status Monitoring information
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: object
            type: array
        "400":
          description: Invalid Parameters
          schema:
            $ref: '#/definitions/models.CommonRes'
        "401":
          description: Validation Authorization
          schema:
            $ref: '#/definitions/models.CommonRes'
        "417":
          description: Nginx Monitoring Info got Exception
          schema:
            $ref: '#/definitions/models.CommonRes'
      summary: Get Nginx Agent Monitoring information
      tags:
      - Monitoring
    post:
      consumes:
      - application/json
      description: Janus Agent and Janus Trun Status Monitoring information register
      parameters:
      - description: BroadCast is close. All Agent and live Stream Api send to broadcast
          close signal
        in: body
        name: BroadCastCloseInfo
        required: true
        schema:
          $ref: '#/definitions/models.MonitoringNginxParams'
      produces:
      - application/json
      responses:
        "200":
          description: Janus Monitoring Info registered
          schema:
            $ref: '#/definitions/models.CommonRes'
        "400":
          description: Invalid Parameters
          schema:
            $ref: '#/definitions/models.CommonRes'
        "401":
          description: Validation Authorization
          schema:
            $ref: '#/definitions/models.CommonRes'
        "417":
          description: Janus Monitoring Info registered Exception
          schema:
            $ref: '#/definitions/models.CommonRes'
      summary: Register Janus Monitoring information
      tags:
      - Monitoring
...
swagger: "2.0"